Treat Source.Tag as a json.Number
If a tag is specified in yaml like:

```
tag: 2
```

It will get turned into json as `{"tag": 2}`. We could try to fix this
at the Go yaml->json level, but I think it's harder than just handling
either here. json.Number can handle a thing that is totally a string
(`{"tag": "foo"}`), or just sorta a string (`{"tag": 2.5}`).